Include some optimizations for actual shirt.
Add a tuneable parameter for some lights to be brighter than others.
This is needed to ounterbalance the effect of lights in the center of
the design having more batting in front of them; in order to appear to
shine with the same intensity, they must be brighter than lights on the
edges of the design.

Comment out all serial debug; this just makes the lights flash
unhelpfully.